+++ gva/data/validator/is_list.py
@@ -1,0 +1,4 @@
+from typing import Any
+
+def is_list(value: Any) -> bool:
+    return type(value).__name__ == 'list'
