+++ gva/data/validator/is_null.py
@@ -1,0 +1,4 @@
+from typing import Any
+
+def is_null(value: Any) -> bool:
+    return not value
