+++ gva/data/validator/other_validator.py
@@ -1,0 +1,4 @@
+from typing import Any
+
+def other_validator(value: Any) -> bool:
+    return True
